requirements.txt
# ---- Core LLM / Agent ----
langchain>=0.2.0
langgraph>=0.0.40
langchain-community>=0.2.0
langchain-openai>=0.1.0
openai>=1.0.0
langchain-huggingface>=0.1.0

# ---- Embeddings / Models ----
sentence-transformers>=2.6.0
transformers>=4.40.0
torch>=2.0.0

# ---- Vector Store ----
faiss-cpu>=1.8.0

# ---- MLOps ----
mlflow>=2.12.0

# ---- API Serving ----
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.6.0

# ---- Utilities ----
python-dotenv>=1.0.0
pytest>=7.0.0
